oops not sure why i changed these

+++ src/boot.h
@@ -2,11 +2,11 @@
 #define _BOOT_H_
 #include <ultra64.h>
 
-u8 *get_csegmentSegmentStart(void);
-u8 *get_cdataSegmentRomStart(void);
-u8 *get_cdataSegmentRomEnd(void);
-u8 *get_rarezipSegmentRomStart(void);
-u8 *get_rarezipSegmentRomEnd(void);
+u32 *get_csegmentSegmentStart(void);
+u32 *get_cdataSegmentRomStart(void);
+u32 *get_cdataSegmentRomEnd(void);
+u32 *get_rarezipSegmentRomStart(void);
+u32 *get_rarezipSegmentRomEnd(void);
 u32 jump_decompressfile(u32 source, u32 target, u32 buffer);
 
 #endif
